+++ RELEASENOTES.md
@@ -1,1 +1,5 @@
 - The script generated from generate-script will now stop on the first command that fails. Previously this script could be renamed between .ps1 and .sh, but with this change it will now only work as a .ps1 script.
+- Adds the ability to migrate ADO repos using the `gh gei` CLI. This overlaps with some of the capabilities of ado2gh, but the `gh gei` will not include all the extra ADO migration capabilities like re-wiring pipelines and boards integration.
+    - `gh gei generate-script` now has an `--ado-source-org` option
+    - `gh gei migrate-repo` now has `--ado-source-org` and `--ado-team-project` options
+- Added `grant-migrator-role` and `revoke-migrator-role` commands to `gh gei`